web_app/backend/requirements.txt
# FastAPI and server dependencies
fastapi>=0.95.0
uvicorn>=0.22.0
python-multipart>=0.0.6
pydantic>=2.0.0

# Authentication
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
PyJWT>=2.6.0

# Environment and configuration
python-dotenv>=1.0.0

# Database
supabase>=1.0.3
psycopg2-binary>=2.9.6

# LlamaIndex and AI dependencies
llama-index>=0.8.0
llama-index-llms-openai>=0.1.0
llama-index-readers-web>=0.1.0
openai>=1.0.0

# Utilities
requests>=2.28.2
typing-extensions>=4.5.0

# Testing
pytest>=7.3.1
pytest-asyncio>=0.21.0
httpx>=0.24.0